+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* Prevented app to crash in case requested app wasn't ready to proceed payment
+* Did some changes in response status because PhonePe was returning **FAILED** instead of **FAILURE** as specified.
+
 ## 1.0.0
 
 * Plugin to implement UPI in Android